build: NetBSD fixes for use of the macro toupper() and tolower()
gcc on NetBSD fails when passing char* to the to{upper,lower} macros.
Cast to a int to work around this.

CR: dalmeida@likewise.com

(pstore: r39716)

+++ pstore/utils/lwps-str.c
@@ -285,7 +285,7 @@ LwpsStrToUpper(
         return;
 
     while (*pszString != '\0') {
-        *pszString = toupper(*pszString);
+        *pszString = toupper((int)*pszString);
         pszString++;
     }
 }
@@ -299,7 +299,7 @@ LwpsStrToLower(
         return;
 
     while (*pszString != '\0') {
-        *pszString = tolower(*pszString);
+        *pszString = tolower((int)*pszString);
         pszString++;
     }
 }
